+++ webpack.config.babel.js
@@ -1,9 +1,5 @@
 import path from 'path';
 import webpack from 'webpack';
-import precss from 'precss';
-import autoprefixer from 'autoprefixer';
-
-const appPath = path.resolve(__dirname, 'src');
 
 const webpackConfig = {
   resolve: {
@@ -19,7 +15,7 @@ const webpackConfig = {
     'prop-types': 'prop-types'
   },
   output: {
-    path: path.join(__dirname, 'dist'), //打包输出目录
+    path: path.join(__dirname, 'build'), //打包输出目录
     filename: '[name].js', //文件名称
     publicPath: './', //生成文件基于上下文路径
     library: ['reactScroll'],
@@ -32,21 +28,19 @@ const webpackConfig = {
         enforce: 'pre',
         test: /\.js$/,
         exclude: /node_modules/,
-        loader: 'eslint-loader'
+        use: {
+          loader: 'eslint-loader',
+          options: {
+            configFile: '.eslintrc',
+            emitError: true, // 验证失败，终止
+          }
+        }
       },
       {
         test: /\.js$/,
         loader: 'babel-loader',
         exclude: /(node_modules)/
       },
-      {
-        test: /\.css$/,
-        loader: 'style-loader!css-loader!postcss-loader?pack=cleaner'
-      },
-      {
-        test: /\.scss$/,
-        loader: 'style-loader!css-loader!postcss-loader?pack=cleaner!sass-loader?outputStyle=expanded'
-      }
     ]
   },
 
@@ -56,25 +50,6 @@ const webpackConfig = {
     new webpack.optimize.AggressiveMergingPlugin(),
     //用来保证编译过程不出错
     new webpack.NoEmitOnErrorsPlugin(),
-    new webpack.LoaderOptionsPlugin({
-      options: {
-        // eslint 配置
-        eslint: {
-          emitError: true, // 验证失败，终止
-          configFile: '.eslintrc'
-        },
-        postcss () {
-          return {
-            defaults: [precss, autoprefixer],
-            cleaner: [autoprefixer({
-              flexbox: 'no-2009',
-              browsers: ['Chrome >= 35', 'Firefox >= 38', 'Edge >= 12',
-                'Explorer >= 8', 'Android >= 4.3', 'iOS >=8', 'Safari >= 8']
-            })]
-          };
-        },
-      }
-    })
   ]
 };